requirements.txt
groq>=0.4.0
python-dotenv>=1.0.0
markdown>=3.4.0
weasyprint>=60.0
flask>=3.0.0

